+++ lib/linux/lnxdatabase.h
@@ -63,7 +63,7 @@ public:
   virtual bool write(const char *label, int entry);
 
   // get the current user's name.
-  virtual void get_user_name(char *buffer, ulong *size);
+  virtual void get_user_name(char *buffer, size_t *size);
 };
 
 // Handy macro to read an int without having to specify the wordsize
