+++ scripts/lnx/osiris_vector.h
@@ -33,7 +33,7 @@ const vector Zero_vector = {0.0f, 0.0f, 0.0f};
 typedef unsigned short angle;
 
 // The basic fixed-point type
-typedef long fix;
+typedef int32_t fix;
 
 #define PI 3.141592654
 
@@ -67,7 +67,7 @@ fix FloatToFixFast(float num);
 //??#define FloatToFix(num) Round((num) * FLOAT_SCALER)
 #define FloatToFix(num) ((fix)((num) * FLOAT_SCALER))
 #define IntToFix(num) ((num) << FIX_SHIFT)
-#define ShortToFix(num) (((long)(num)) << FIX_SHIFT)
+#define ShortToFix(num) (((int32_t)(num)) << FIX_SHIFT)
 #define FixToFloat(num) (((float)(num)) / FLOAT_SCALER)
 #define FixToInt(num) ((num) >> FIX_SHIFT)
 #define FixToShort(num) ((short)((num) >> FIX_SHIFT))
