+++ sndlib/ddsoundload.cpp
@@ -122,12 +122,12 @@ char SoundLoadWaveFile(const char *filename, float percent_volume, int sound_fil
 
   char format_type[80];       // ASCII name of format type
   unsigned short fmttag = 0;  // Numerical format type
-  unsigned long ckid;         // Current chunk's ID
-  unsigned long cksize;       // Current chunk's size in bytes
-  unsigned long filesize;     // Size of the sound file
-  unsigned long nextseek = 0; // Location of the next seek
+  uint32_t ckid;         // Current chunk's ID
+  uint32_t cksize;       // Current chunk's size in bytes
+  uint32_t filesize;     // Size of the sound file
+  uint32_t nextseek = 0; // Location of the next seek
 
-  unsigned long aligned_size; // Sound files are aligned to SOUND_FILE_SAMPLE_ALIGNMENT samples
+  uint32_t aligned_size; // Sound files are aligned to SOUND_FILE_SAMPLE_ALIGNMENT samples
 
   // Sound format information
   int samples_per_second;
@@ -135,7 +135,7 @@ char SoundLoadWaveFile(const char *filename, float percent_volume, int sound_fil
   short number_channels;
 
   // Used to read temporary long values
-  unsigned long temp_long;
+  uint32_t temp_long;
 
   // Flags for if we previously read data or a format
   char f_data, f_fmt = 0;
@@ -169,7 +169,7 @@ char SoundLoadWaveFile(const char *filename, float percent_volume, int sound_fil
   paged_in_count += cfilelength(cfptr);
   paged_in_num++;
   // Make sure that it is a RIFF format
-  temp_long = (unsigned long)cf_ReadInt(cfptr);
+  temp_long = (uint32_t)cf_ReadInt(cfptr);
   if (temp_long != 0x46464952) {
     mprintf((0, "SOUND LOADER: %s is not a RIFF format file\n", filename));
     goto error_state;
@@ -180,7 +180,7 @@ char SoundLoadWaveFile(const char *filename, float percent_volume, int sound_fil
   filesize += cftell(cfptr);
 
   // Make sure it is a wave file
-  temp_long = (unsigned long)cf_ReadInt(cfptr);
+  temp_long = (uint32_t)cf_ReadInt(cfptr);
   if (temp_long != 0x45564157) {
     mprintf((0, "SOUND LOADER:  %s is not a WAVE file\n", filename));
     goto error_state;
